修改首页商品查询接口

+++ mall-service/src/main/java/com/xingyun/bbc/mall/model/vo/UserWalletDetailVo.java
@@ -37,7 +37,7 @@ public class UserWalletDetailVo {
 
     /** 余额 */
     @ApiModelProperty("余额")
-    private Long fbalance;
+    private BigDecimal fbalance;
 
 
     /** 修改时间 */
